Day 3

+++ app/src/main/java/com/example/premierleaguenewsfilter/data/DatabasePlayerItem.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 
 @Entity
 data class DatabasePlayerItem(
-    @PrimaryKey(autoGenerate = true) val uid: Int,
+    @PrimaryKey(autoGenerate = true) val uid: Long,
     @ColumnInfo(name = "first_name") val firstName: String,
     @ColumnInfo(name = "last_name") val lastName: String,
     @ColumnInfo(name = "club") val club: String,
